requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.2.0
typer>=0.9.0
rich>=13.7.0
structlog>=23.2.0

# Essential LangChain (minimal set)
langchain-core>=0.2.0
langchain>=0.2.0
langgraph>=0.1.0
langsmith>=0.1.0

# LLM Provider Support
langchain-aws>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.2
pyyaml>=6.0.1

# HTTP and API
httpx>=0.25.2
requests>=2.31.0

# Async and Concurrency
aiofiles>=23.2.1

# Database dependencies for PostgreSQL memory store
asyncpg>=0.29.0
psycopg2-binary>=2.9.0



# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.11.0
flake8>=6.1.0

# MCP (Model Context Protocol)
mcp>=1.0.0